bug fixed

+++ RFID_Station_control_android/App2/Activity_status.cs
@@ -162,7 +162,7 @@ namespace RfidStationControl
             {
                 if (!chipsCheckedText.HasFocus)
                 {
-                    byte.TryParse(chipsCheckedText.Text, out var n);
+                    ushort.TryParse(chipsCheckedText.Text, out var n);
                     chipsCheckedText.Text = n.ToString();
                     StatusPageState.CheckedChipsNumber = n;
                 }
